refactor: remove deprecated GeneratedVaadinDialog
