+++ vaadin-dialog-flow-parent/vaadin-dialog-flow/pom.xml
@@ -31,6 +31,11 @@
         </dependency>
         <dependency>
             <groupId>com.vaadin</groupId>
+            <artifactId>vaadin-flow-components-base</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.vaadin</groupId>
             <artifactId>vaadin-renderer-flow</artifactId>
             <version>${project.version}</version>
         </dependency>
